Add files via upload
This project is a part of research about News related to different crimes in Pakistan. This script is created to automatically extract the News of One Year from the Archive of Dawn News paper. This script scrape the HTML pages and get the complete contents of a News and write this to a CSV file